[codex-api] isolate websocket auth tests [ci changed_files]
